my_vector::insert 已正确使用临时副本 tmp，自引用插入问题已修复。 my_string 的移动构造和移动赋值都会将源对象恢复为空字符串。 my_utils.hpp 中的移动交换和比较器 const 问题已处理。

+++ 数据结构与算法/tests/test4/calc_xor_sum.hpp
@@ -8,7 +8,7 @@ inline void calc_xor_sum(my_list<int> &lst)
 {
     int ans = 0;
     my_list<int>::iterator it = lst.begin();
-    for (int i = 0; i < lst.size(); i++)
+    for (size_t i = 0; i < lst.size(); i++)
     {
         ans += (i ^ *it);
         it++;
